+++ app/helper/aspect_image.py
@@ -30,7 +30,22 @@ def get_city_image(selected_city_long_name: str) -> AspectImage:
         selected_city_long_name.strip("_small")
     )
 
-    if selected_city in ["Hallein", "Dürrnberg"]:
+    if selected_city == "Hallein":
+        if "small" in selected_city_long_name:
+            return AspectImage(
+                source=get_logo_file_path(selected_city),
+                # values are highly specific to used PNG
+                size_hint=(1.05, 1.05),
+                pos_hint={"x": -0.05, "y": -0.05},
+            )
+        else:
+            return AspectImage(
+                source=get_logo_file_path(selected_city),
+                # values are highly specific to used PNG
+                size_hint=(None, 0.21),
+                pos_hint={"x": 0, "y": 0.79},
+            )
+    elif selected_city == "Dürrnberg":
         if "small" in selected_city_long_name:
             return AspectImage(
                 source=get_logo_file_path(selected_city),