+++ app/screens/competition_menu.py
@@ -1,8 +1,6 @@
 from kivy.app import App
 from kivy.uix.button import Button
-from kivy.uix.switch import Switch
 from kivy.uix.label import Label
-from kivy.uix.layout import Layout
 from kivy.uix.screenmanager import Screen
 
 from typing import cast