+++ app/screens/login_screen.py
@@ -1,11 +1,12 @@
+from kivy.uix.label import Label
+from kivy.uix.togglebutton import ToggleButton
 from kivy.uix.screenmanager import Screen
 from kivy.clock import Clock
-from kivy.uix.togglebutton import ToggleButton
 
 from helper.functions import change_screen_to
 from helper.file_handling import update_config, map_selected_city_2long_name
 from helper.aspect_image import get_city_image
-from helper.settings import Strings
+from helper.settings import Strings, Info_Text
 
 
 strings = Strings()
@@ -14,11 +15,8 @@ strings = Strings()
 class Login(Screen):
     def __init__(self, **kwargs):
         super(Login, self).__init__(**kwargs)
-
         self.ids.login_label.text = strings.LABEL_STR_LOGIN
-
         self.ids.store_selection_button.text = strings.BUTTON_STR_STORE_SELECTION
-
         self.ids.store_selection_button.disabled = True
 
     def on_pre_enter(self):
@@ -38,8 +36,8 @@ class Login(Screen):
 
         for city, state in self.cities:
             btn = ToggleButton(
-                text=f"{city}{' '*3}[size=30]{state}[/size]",
-                markup=True,  # Enable markup for custom text positioning
+                text=f"{city}\n[size=35]{state}[/size]",
+                markup=True,
                 font_size="32sp",
                 group="city",
                 allow_no_selection=True,
@@ -49,14 +47,25 @@ class Login(Screen):
             btn.bind(on_release=self.on_city_toggle)
             self.ids.city_layout.add_widget(btn)
 
+        lbl = Label(
+            text=Info_Text().TEXT,
+            markup=True,
+            font_size="20sp",
+            size_hint_y=None,
+            height="150dp",
+        )
+        self.ids.city_layout.add_widget(lbl)
+
     def on_city_toggle(self, instance):
         if instance.state == "down":
             self.selected_button = instance.text
 
-            # deactivate others
-            for btn in self.ids.city_layout.children:
-                if btn != instance and btn.state == "down":
-                    btn.state = "normal"
+            # deactivate others ToggleButtons
+            for child in self.ids.city_layout.children:
+                # ignore Label
+                if isinstance(child, ToggleButton):
+                    if child != instance and child.state == "down":
+                        child.state = "normal"
 
             self.ids.store_selection_button.disabled = False
 
@@ -65,7 +74,7 @@ class Login(Screen):
             self.ids.store_selection_button.disabled = True
 
     def store_selection(self):
-        state = self.selected_button.split("   ")[1]
+        state = self.selected_button.split("\n")[1]
         self.selected_state = state[9:-7]
         self.selected_city = map_selected_city_2long_name(self.selected_button)