+++ app/screens/score_screen.py
@@ -1,4 +1,3 @@
-from kivy.uix.label import Label
 from kivy.uix.screenmanager import Screen
 
 from typing import cast
@@ -10,11 +9,6 @@ about_text = About_Text().TEXT
 strings = Strings()
 
 
-class Info_Screen(Screen):
-    def __init__(self, **kwargs):
-        super(Info_Screen, self).__init__(**kwargs)
-        self.header_label = cast(Label, self.header_label)
-        self.header_label.text = strings.BUTTON_STR_INFO
-
+class Punkte_Screen(Screen):
     def go_back(self, *args) -> None:
         change_screen_to("start_menu")